requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SeaSurf==1.1.1
Flask-Talisman==1.1.0
Flask-LDAP3-Login==0.9.17
Flask-Authorize==0.2.7
psycopg2-binary==2.9.7
mysql-connector-python==8.1.0
celery==5.3.6
redis==4.6.0
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.0.0
WTForms==3.0.1
ldap3==2.9.1
gunicorn==21.2.0
Werkzeug==2.3.7
sqlalchemy>=2.0.0